Update World boundaries in FSD w/ cryostat GDML
Increaed the world volume boundaries by 10x in the FSD with cryostat
geometry. This somehow fixed a segfault that was affecting up to 33% of
CORSIKA files when being processed with edep-sim/Geant4.

This should not change the results as the world volume is vacuum.

Original bounds are still in the file with a comment.