Modify convert script event numbering for cosmic sim
Modified the convert_edepsim script to renumber the event IDs for cosmic
events to be consecutive. The vertex ID is unchanged and will match with
the edep-sim event ID.

This is to work better with the larnd-sim simulation properties setting
of `max_events_per_file`, which needs the maximum event ID value in the
file (or something greater).

Also updated FSD workflow yaml used in run_ndlar_flow_fsd.sh